refactor: vars generation variable (#29)
