+++ beginner/variables/base.go
@@ -1,0 +1,6 @@
+package variables
+
+var (
+	name, location string
+	age            int
+)
