+++ beginner/variables/functions.go
@@ -1,0 +1,5 @@
+package variables
+
+var anonym = func() string {
+	return "Itachi Uchiha"
+}
