--- beginner/variables/variables_functions.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-	"fmt"
-)
-
-var anonima = func() {
-	fmt.Printf("Itachi Uchiha")
-}
-
-func main() {
-	anonima()
-}
