--- beginner/variables/variables_initializers.go
@@ -1,21 +1,0 @@
-package main
-
-import (
-	"fmt"
-)
-
-var (
-	name     = "Luis Mayta"
-	age      = 30
-	location = "Lima - Peru"
-	alias    = "Itachi uchiha"
-)
-
-func main() {
-	fmt.Printf("%s have %d from %s alias %s",
-		name,
-		age,
-		location,
-		alias,
-	)
-}
